Fixing bigint > i64

+++ src/common.rs
@@ -76,7 +76,7 @@ pub struct Real {
 #[derive(Clone, Debug, Eq, Hash, PartialEq, Serialize, Deserialize)]
 pub enum Literal {
     Null,
-    Integer(i64),
+    Integer(i128),
     FixedPoint(Real),
     String(String),
     Blob(Vec<u8>),
@@ -86,8 +86,8 @@ pub enum Literal {
     Placeholder,
 }
 
-impl From<i64> for Literal {
-    fn from(i: i64) -> Self {
+impl From<i128> for Literal {
+    fn from(i: i128) -> Self {
         Literal::Integer(i)
     }
 }
@@ -820,7 +820,7 @@ named!(pub integer_literal<&[u8], Literal>,
         sign: opt!(tag!("-")) >>
         val: digit >>
         ({
-            let mut intval = i64::from_str(str::from_utf8(val).unwrap()).unwrap();
+            let mut intval = i128::from_str(str::from_utf8(val).unwrap()).unwrap();
             if sign.is_some() {
                 intval *= -1;
             }
