+++ src/create.rs
@@ -281,7 +281,7 @@ named!(pub column_constraint<&[u8], Option<ColumnConstraint>>,
                               })
                     ))
                   | do_parse!(d: map_res!(digit, str::from_utf8) >> (
-                        Literal::Integer(i64::from_str(d).unwrap())
+                        Literal::Integer(i128::from_str(d).unwrap())
                     ))
                   | do_parse!(tag!("''") >> (Literal::String(String::from(""))))
                   | do_parse!(tag_no_case!("null") >> (Literal::Null))
